extras/grpc_client/requirements.txt
pandas
numpy
grpcio
grpcio-tools
protobuf 